Moved test file into correct directory and removed requirejs-ness
Fixes the test failures introduced in https://github.com/alphagov/spotlight/pull/737
